Added:Functionality for Validations and Exception Handling

+++ src/main/java/com/capgemini/employeepayrollapp/repositories/IEmployeeRepository.java
@@ -2,6 +2,6 @@ package com.capgemini.employeepayrollapp.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.capgemini.employeepayrollapp.model.Employee;
 
-public interface IEmployeeRepository extends JpaRepository<Employee,Long>{
+public interface IEmployeeRepository extends JpaRepository<Employee,Integer>{
 
 }
